fix import errors + move util functions to seperate file
