+++ app/routes/fingerprint.py
@@ -8,8 +8,8 @@ import requests
 import time
 
 from app import models
-from app.app import (
-    db,
+from app.app import db
+from app.util import (
     lockbot_request,
     mattermost_doorkeeper_message,
     mattermost_response,
